Refactor review screen
