+++ apps/web/src/screens/review/ReviewHardReminderDialog.tsx
@@ -1,7 +1,7 @@
 import type { ReactElement } from "react";
 import { useI18n } from "../../i18n";
 
-type ReviewHardReminderDialogProps = Readonly<{
+export type ReviewHardReminderDialogProps = Readonly<{
   isOpen: boolean;
   onDismiss: () => void;
 }>;
